requirements.txt
dnspython
graphviz
rich
